test: community page filter and ordering
